Move `to_snake_case` to string utils
