--- lib/streamlit/case_converters.py
@@ -1,30 +1,0 @@
-# Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022-2024)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import annotations
-
-import re
-
-
-def to_snake_case(camel_case_str: str) -> str:
-    """Converts UpperCamelCase and lowerCamelCase to snake_case.
-
-    Examples
-    --------
-        fooBar -> foo_bar
-        BazBang -> baz_bang
-
-    """
-    s1 = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", camel_case_str)
-    return re.sub("([a-z0-9])([A-Z])", r"\1_\2", s1).lower()
