+++ lib/streamlit/runtime/app_session.py
@@ -21,8 +21,8 @@ from enum import Enum
 from typing import TYPE_CHECKING, Callable, Final
 
 import streamlit.elements.exception as exception_utils
+from lib.streamlit.string_util import to_snake_case
 from streamlit import config, runtime
-from streamlit.case_converters import to_snake_case
 from streamlit.logger import get_logger
 from streamlit.proto.ClientState_pb2 import ClientState
 from streamlit.proto.Common_pb2 import FileURLs, FileURLsRequest
